Finished reads of a single key

+++ examples/src/tiledb_dense_read_async.cc
@@ -92,8 +92,9 @@ int main() {
   printf("-----------------------------------------\n");
   for (uint64_t i = 0; i < result_num; ++i) {
     printf("%3d", buffer_a1[i]);
-    size_t var_size = (i != result_num - 1) ? buffer_a2[i + 1] - buffer_a2[i] :
-                                              buffer_sizes[2] - buffer_a2[i];
+    uint64_t var_size = (i != result_num - 1) ?
+                            buffer_a2[i + 1] - buffer_a2[i] :
+                            buffer_sizes[2] - buffer_a2[i];
     printf("\t %4.*s", int(var_size), &buffer_var_a2[buffer_a2[i]]);
     printf("\t\t (%5.1f, %5.1f)\n", buffer_a3[2 * i], buffer_a3[2 * i + 1]);
   }
